Files necessary for running the markdown
The files "regression_data_v4.csv", "Variables_v9_raw.csv" and the markdown document need to be in the same folder. Put all files beginning with "cb_2015_us_county_500k" in a folder called "Geospatial Data" and put that folder in the folder where you saved the csvs and the markdown document. Set the working directory in R to that folder. Run the model.